Add namespaces for Compression, IO, and Serialization
