+++ Chickensoft.SaveFileBuilder/src/IO/FileIO.cs
@@ -1,4 +1,4 @@
-namespace Chickensoft.SaveFileBuilder;
+namespace Chickensoft.SaveFileBuilder.IO;
 
 using System.IO;
